Deleted SimpleVector - Point already provides this.

+++ Source/Characters/Character.h
@@ -23,9 +23,9 @@ class Character
   void wake(ASGE::Renderer* passed_renderer);
   void updatePosition(double delta_time);
 
-  void setSpawnPositionX(int x_pos);
-  void setSpawnPositionY(int y_pos);
-  void setSpawnPosition(int x_pos, int y_pos);
+  void setSpawnPositionX(float x_pos);
+  void setSpawnPositionY(float y_pos);
+  void setSpawnPosition(float x_pos, float y_pos);
   void setVisible(bool is_visible);
   void setDimensions(int new_width, int new_height);
   void setSpeed(int speed);
